+++ packages/@headlessui-react/src/components/dialog-backdrop/dialog-backdrop.tsx
@@ -1,3 +1,0 @@
-// Next.js barrel file improvements (GENERATED FILE)
-export type * from '../dialog/dialog'
-export { DialogBackdrop } from '../dialog/dialog'